Adds copying_task_factory and tests
The copying_task_factory does not share the context
when creating new tasks and uses a copy of it.
This is to allow created tasks to modify their context
without mirroring the change into every other task, but
whilst still passing along a copy of their own context
to their children.